Migrate directory checks to Save and SaveAs, instead of saveToFile
By migrating these checks, we can have Save handle the following three cases:

* A buffer has an empty path, and needs to call SaveAs.
* A buffer has a path that points to a directory, and needs to call SaveAs to get a new filename.
* A buffer has a path that is a file, and needs to cause saveToFile.

The code that was originally added to saveToFile has now been migrated to SaveAs. Instead of SaveAs prompting for the filename with a blank, it now uses what that buffer's path is as a basis. This should either be blank or the directory path. By migrating our check here, we can also ask about overwriting files, which solves the remaining gap in the previous approach.