Default hlsearch colors
Like in Vim, if hlsearch is not defined in the colorscheme, use the
reverse default colors (same as for selection).
This is the most universal solution (the least dependent on possible
terminal limitations) but it's not ideal, since hlsearch is
indistinguishable from selection (considering that selection is still
not defined in any of the built-in colorschemes).