Markdown: highlight after brace. Solves (#2290)
Currently highlihting of ** **, __ __, ~~ ~~, ` ` syntax is allowed
only first part of the markup is after space character or in the
start of the string. It was done to preven highliting if the markup
symbol (for example undercore in variable name) will occur inside the
word. It means that (** **) won't be highlighted even though it usually
perceived as allowed by people

Signed-off-by: Yevhen Babiichuk (DustDFG) <dfgdust@gmail.com>