TryBindKey: don't write to bindings.json unless overwrite is true
Change the (documented) behavior of TryBindKey(): unless the overwrite
argument is true, never write to bindings.json, even if the new binding
has been successfully applied.

TryBindKey() with overwrite=false is mainly used by plugins for adding
their own default keybindings. It is not really necessary for them to
write those default bindings to bindings.json: even if the binding is
not saved in bindings.json, it will be still applied by the plugin
calling TryBindKey() every time micro starts.

This fixes undesirable behavior: unless there is already a binding for
the given key in bindings.json, a plugin forcibly rewrites
bindings.json, in particular removing any user's comments from it.
With this change, bindings.json will be only rewritten by micro in the
case when the user runs `bind` or `unbind` command, not automatically by
plugins (unless there is a plugin calling TryBindKey() with
overwrite=true, which is unlikely).